refactor: move popup state things into separate module
popup.tsx is about as large as I'm comfortable with, and I need these to
be separate to avoid circular dependencies when importing separate tsx
modules into popup.
